# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type TaskStatus = typeof TaskStatus[keyof typeof TaskStatus];


export const TaskStatus = {
  pending: 'pending',
  in_progress: 'in_progress',
  done: 'done',
} as const;

export type TaskPriority = typeof TaskPriority[keyof typeof TaskPriority];


export const TaskPriority = {
  low: 'low',
  medium: 'medium',
  high: 'high',
} as const;

export interface Task {
  id: number;
  title: string;
  /** @nullable */
  description?: string | null;
  status: TaskStatus;
  priority: TaskPriority;
  /** @nullable */
  dueDate?: string | null;
  createdAt: string;
  updatedAt?: string;
}

export type TaskInputStatus = typeof TaskInputStatus[keyof typeof TaskInputStatus];


export const TaskInputStatus = {
  pending: 'pending',
  in_progress: 'in_progress',
  done: 'done',
} as const;

export type TaskInputPriority = typeof TaskInputPriority[keyof typeof TaskInputPriority];


export const TaskInputPriority = {
  low: 'low',
  medium: 'medium',
  high: 'high',
} as const;

export interface TaskInput {
  /** @minLength 1 */
  title: string;
  description?: string;
  status?: TaskInputStatus;
  priority?: TaskInputPriority;
  dueDate?: string;
}

export type TaskUpdateStatus = typeof TaskUpdateStatus[keyof typeof TaskUpdateStatus];


export const TaskUpdateStatus = {
  pending: 'pending',
  in_progress: 'in_progress',
  done: 'done',
} as const;

export type TaskUpdatePriority = typeof TaskUpdatePriority[keyof typeof TaskUpdatePriority];


export const TaskUpdatePriority = {
  low: 'low',
  medium: 'medium',
  high: 'high',
} as const;

export interface TaskUpdate {
  /** @minLength 1 */
  title?: string;
  description?: string;
  status?: TaskUpdateStatus;
  priority?: TaskUpdatePriority;
  /** @nullable */
  dueDate?: string | null;
}

export interface Note {
  id: number;
  title: string;
  content: string;
  pinned?: boolean;
  createdAt: string;
  updatedAt?: string;
}

export interface NoteInput {
  /** @minLength 1 */
  title: string;
  content?: string;
  pinned?: boolean;
}

export interface NoteUpdate {
  /** @minLength 1 */
  title?: string;
  content?: string;
  pinned?: boolean;
}

export interface Reminder {
  id: number;
  title: string;
  remindAt: string;
  done: boolean;
  createdAt: string;
}

export interface ReminderInput {
  /** @minLength 1 */
  title: string;
  remindAt: string;
}

export interface ReminderUpdate {
  title?: string;
  remindAt?: string;
  done?: boolean;
}

export type ChatMessageRole = typeof ChatMessageRole[keyof typeof ChatMessageRole];


export const ChatMessageRole = {
  user: 'user',
  assistant: 'assistant',
} as const;

export interface ChatMessage {
  id: number;
  role: ChatMessageRole;
  content: string;
  createdAt: string;
}

export interface ChatMessageInput {
  /** @minLength 1 */
  content: string;
}

export interface DashboardStats {
  totalTasks: number;
  doneTasks: number;
  pendingTasks: number;
  totalNotes: number;
  totalReminders: number;
  upcomingReminders: number;
}

export type ActivityItemType = typeof ActivityItemType[keyof typeof ActivityItemType];


export const ActivityItemType = {
  task: 'task',
  note: 'note',
  reminder: 'reminder',
  chat: 'chat',
} as const;

export interface ActivityItem {
  id: string;
  type: ActivityItemType;
  title: string;
  createdAt: string;
}

export type ListTasksParams = {
status?: ListTasksStatus;
priority?: ListTasksPriority;
};

export type ListTasksStatus = typeof ListTasksStatus[keyof typeof ListTasksStatus];


export const ListTasksStatus = {
  pending: 'pending',
  in_progress: 'in_progress',
  done: 'done',
} as const;

export type ListTasksPriority = typeof ListTasksPriority[keyof typeof ListTasksPriority];


export const ListTasksPriority = {
  low: 'low',
  medium: 'medium',
  high: 'high',
} as const;

export type ListRemindersParams = {
upcoming?: boolean;
};

export type ListChatMessagesParams = {
limit?: number;
};