# lib/api-zod/src/generated/types/reminder.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Reminder {
  id: number;
  title: string;
  remindAt: string;
  done: boolean;
  createdAt: string;
}