# lib/api-zod/src/generated/types/task.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { TaskPriority } from './taskPriority';
import type { TaskStatus } from './taskStatus';

export interface Task {
  id: number;
  title: string;
  /** @nullable */
  description?: string | null;
  status: TaskStatus;
  priority: TaskPriority;
  /** @nullable */
  dueDate?: string | null;
  createdAt: string;
  updatedAt?: string;
}